feat: scaffold cli with server cmd
