+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod event;
 pub mod persist;
 pub mod query;
+pub mod server;
 
 pub const LYNX_FORMAT_HEADER: &str = "X-Lynx-Format";
